requirements.txt
openai==0.27.0
streamlit==1.18.1
python-dotenv==1.0.0
asyncio==3.4.3
